refine the code (#143)
